backend/requirements.txt
flask
requests
deemix
packaging
click
setproctitle
wheel
PlexAPI
tqdm
Unidecode
eventlet
flask-socketio
watchdog
pillow